requirements.txt
streamlit
google-genai
requests
python-dotenv
pydantic
